+++ src/containers/ListingPage/ListingPageCoverPhoto.js
@@ -279,8 +279,13 @@ export const ListingPageComponent = props => {
       }
     : {};
   const currentStock = currentListing.currentStock?.attributes?.quantity || 0;
-  const schemaAvailability =
-    currentStock > 0 ? 'https://schema.org/InStock' : 'https://schema.org/OutOfStock';
+  const schemaAvailability = !currentListing.currentStock
+    ? null
+    : currentStock > 0
+    ? 'https://schema.org/InStock'
+    : 'https://schema.org/OutOfStock';
+
+  const availabilityMaybe = schemaAvailability ? { availability: schemaAvailability } : {};
 
   const handleViewPhotosClick = e => {
     // Stop event from bubbling up to prevent image click handler
@@ -307,7 +312,7 @@ export const ListingPageComponent = props => {
           '@type': 'Offer',
           url: productURL,
           ...schemaPriceMaybe,
-          availability: schemaAvailability,
+          ...availabilityMaybe,
         },
       }}
     >